// src/contexts/Farms/types.ts
import { Contract } from 'web3-eth-contract'

export interface Farm {
  pid: number
  base: string
  name: string
  lpToken: string
  lpTokenAddress: string
  lpContract: Contract
  tokenAddress: string
  earnToken: string
  earnTokenAddress: string
  icon: React.ReactNode
  id: string
  tokenSymbol: string
}

export interface FarmsContext {
  farms: Farm[]
  unharvested: number
}